move stats functions to stats module
